+++ src/app/components/auth/auth.component.html
@@ -93,12 +93,17 @@
                       J'accepte les conditions d'utilisation <a href="" disabled>CDU</a>
                     </label>
                   </div> -->
-                  <div class="alert alert-warning" role="alert" *ngIf="error">
-                    {{error}}
+
+                  <div class="alert alert-success" role="alert" *ngIf="isSuccess">
+                    Enregistrement réussie !
                   </div>
 
                   <div class="d-flex justify-content-center mx-4 mb-3 mb-lg-4">
                     <button type="submit" class="btn btn-dark btn-lg" [disabled]="!signinForm.valid">Valider</button>
+                  </div>
+
+                  <div class="alert alert-warning" role="alert" *ngIf="error">
+                    {{error}}
                   </div>
                 </form>
