+++ src/app/shared/services/person.service.ts
@@ -2,7 +2,7 @@ import { Injectable, OnDestroy } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { environment } from 'src/environments/environment';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { Person } from '../interfaces/person';
+import { Person } from '../models/person';
 
 @Injectable({
   providedIn: 'root'
